Receive an ack when sending a filter-setting msg
Users reported the some dlt messages could not be caught for DLTBroker.
The patch adds an ack to make sure the filter is set properly. In normal
condition, the filter is set in very short time (<2ms). If the DLTClient
does not establish the connection, the action could be delayed. The
patch is more like a debug patch not a function patch.

If there is any error happens after the patch, we have more logs and
informations for the problem.